+++ modules/native_widgets.lua
@@ -198,41 +198,6 @@ local WIDGETS = {
             reset = "EZO_HUD_OPTION_NATIVE_COMBAT_TIPS_RESET",
             resetTooltip = "EZO_HUD_OPTION_NATIVE_COMBAT_TIPS_RESET_TOOLTIP",
         }
-    },
-    {
-        id = "nativeGroupSearchIndicator",
-        controlName = "ZO_ActivityTracker",
-        fallbackAnchor = { RIGHT, GuiRoot, RIGHT, 0, 10 },
-        minScale = 0.75,
-        maxScale = 1.25,
-        onPreviewOpen = function(self, control)
-            if control then
-                control:SetHidden(false)
-                control:SetAlpha(1)
-                local backdrop = GetOrCreatePreviewBackdrop(control, self)
-                if backdrop then backdrop:SetHidden(false) end
-            end
-        end,
-        onPreviewClose = function(self, control)
-            if control then
-                local backdrop = GetOrCreatePreviewBackdrop(control, self)
-                if backdrop then backdrop:SetHidden(true) end
-            end
-        end,
-        stringIds = {
-            header = "EZO_HUD_OPTION_NATIVE_GROUP_SEARCH_INDICATOR",
-            headerTooltip = "EZO_HUD_OPTION_NATIVE_GROUP_SEARCH_INDICATOR_HEADER_TOOLTIP",
-            enable = "EZO_HUD_OPTION_NATIVE_GROUP_SEARCH_INDICATOR_ENABLE",
-            enableTooltip = "EZO_HUD_OPTION_NATIVE_GROUP_SEARCH_INDICATOR_ENABLE_TOOLTIP",
-            offsetX = "EZO_HUD_OPTION_NATIVE_GROUP_SEARCH_INDICATOR_OFFSET_X",
-            offsetXTooltip = "EZO_HUD_OPTION_NATIVE_GROUP_SEARCH_INDICATOR_OFFSET_X_TOOLTIP",
-            offsetY = "EZO_HUD_OPTION_NATIVE_GROUP_SEARCH_INDICATOR_OFFSET_Y",
-            offsetYTooltip = "EZO_HUD_OPTION_NATIVE_GROUP_SEARCH_INDICATOR_OFFSET_Y_TOOLTIP",
-            scale = "EZO_HUD_OPTION_NATIVE_GROUP_SEARCH_INDICATOR_SCALE",
-            scaleTooltip = "EZO_HUD_OPTION_NATIVE_GROUP_SEARCH_INDICATOR_SCALE_TOOLTIP",
-            reset = "EZO_HUD_OPTION_NATIVE_GROUP_SEARCH_INDICATOR_RESET",
-            resetTooltip = "EZO_HUD_OPTION_NATIVE_GROUP_SEARCH_INDICATOR_RESET_TOOLTIP",
-        }
     }
 }
 
@@ -440,17 +405,6 @@ function EZO_HUD:InitializeNativeWidgets()
                 self:ApplyAllNativeWidgetLayouts()
             end
         )
-    end
-
-    if ZO_ACTIVITY_FINDER_ROOT_MANAGER and ZO_ACTIVITY_FINDER_ROOT_MANAGER.RegisterCallback then
-        ZO_ACTIVITY_FINDER_ROOT_MANAGER:RegisterCallback("OnActivityFinderStatusUpdate", function()
-            self:ApplyNativeWidgetLayout("nativeGroupSearchIndicator")
-            if zo_callLater then
-                zo_callLater(function()
-                    self:ApplyNativeWidgetLayout("nativeGroupSearchIndicator")
-                end, 50)
-            end
-        end)
     end
 
     -- Force Keyboard Loot History to be used universally